requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.1
beautifulsoup4==4.6.0
billiard==3.3.0.23
bleach==2.1.4
bs4==0.0.1
builtwith==1.3.3
celery==3.1.26.post2
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
Django==1.11.12
django-celery==3.2.2
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
django-tinymce==2.7.0
djangorestframework==3.8.2
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.6
ioloop==0.1a0
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.8.0
ipython-genutils==0.2.0
jieba==0.39
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
kombu==3.0.37
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
MySQL-python==1.2.5
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.14.1
pandocfilters==1.4.2
pathlib2==2.3.2
pdfminer==20110515
pickleshare==0.7.4
Pillow==5.1.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
Pygments==2.2.0
pymssql==2.1.3
PyMySQL==0.8.0
pyparsing==2.2.0
pyquery==1.4.0
python-dateutil==2.7.3
python-whois==0.7.0
pytz==2018.4
pywinpty==0.5.4
pyzmq==17.1.2
requests==2.18.4
scandir==1.9.0
scipy==1.1.0
selenium==2.48.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
urllib3==1.22
virtualenv==15.2.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
win-unicode-console==0.5
xadmin==0.6.1
